--- config/default.json
@@ -1,39 +1,0 @@
-{
-    "grunt": {
-        "ports": {
-            "server":     1337,
-            "karma":      4669,
-            "livereload": 46692
-        },
-        "folders": {
-            "config":   "config",
-            "client":   "client",
-            "dist":     "dist",
-            "app":      "lib/",
-            "assets":   "",
-            "tasks":    "./tasks",
-            "test":     "./test",
-            "reports":  "./reports",
-            "specs":    "jasmine/specs",
-            "coverage": "coverage",
-            "docs":     "docs"
-        },
-        "files": {
-            "config": {
-                "csslint": "./config/.csslintrc",
-                "eslint":  "./config/.eslintrc.js",
-                "karma":   "./config/karma.conf.js"
-            },
-            "index":        "index.html",
-            "scripts":      "**/*.js",
-            "mainScript":   "main.js",
-            "configScript": "config.js",
-            "models":       "models/**/*.js",
-            "views":        "views/**/*.js",
-            "controllers":  "controllers/**/*.js",
-            "fonts":        "fonts/*.{ttf,woff,eot,svg}",
-            "images":       "images/**/*.{png,jpg,gif,svg}",
-            "templates":    "templates/**/*.hbs"
-        }
-    }
-}
