+++ lib/atmosphere.js
@@ -178,7 +178,7 @@ function temperature(i) {
  * @returns {number} rate (in K/km)
 **/
 function a(i) {
-    var a = [/* eslint-disable no-magic-numbers */
+    var lapseRates = [/* eslint-disable no-magic-numbers */
         -6.5,
         0,
         1,
@@ -200,8 +200,8 @@ function a(i) {
         1.7,
         1.1,
         0
-    ]/* eslint-enable no-magic-numbers */
-    .map(function(a) {return a === 0 ? a : (a / 1000);});
+    ];/* eslint-enable no-magic-numbers */
+    var a = lapseRates.map(function(a) {return a === 0 ? a : (a / 1000);});
     return a[i];
 }
 /**
