+++ lib/index.js
@@ -1,13 +1,8 @@
-/* istanbul ignore next */
-    'use strict';
+'use strict';
 
 require('./polyfill');
-var math       = require('./math');
-var geodetic   = require('./geodetic');
-var atmosphere = require('./atmosphere');
+const math       = require('./math');
+const geodetic   = require('./geodetic');
+const atmosphere = require('./atmosphere');
 
-module.exports = {
-    math,
-    geodetic: geodetic,
-    atmosphere: atmosphere
-};
+module.exports = {math, geodetic, atmosphere};
