+++ package.json
@@ -17,54 +17,30 @@
     "lodash": "^4.17.4"
   },
   "devDependencies": {
-    "aliasify": "^2.1.0",
     "babel-cli": "^6.24.1",
     "babel-preset-babili": "^0.1.4",
     "babel-preset-env": "^1.6.0",
     "babelify": "^7.3.0",
     "browserify": "^14.3.0",
-    "chalk": "^1.1.3",
-    "config": "^1.14.0",
     "deamdify": "^0.3.0",
     "docdash": "^0.4.0",
+    "eslint": "^4.2.0",
     "eslint-config-omaha-prime-grade": "^3.0.2",
-    "eslint-plugin-backbone": "^2.0.1",
     "eslint-plugin-compat": "^1.0.1",
     "fs-promise": "^2.0.2",
     "globby": "^6.0.0",
     "grunt": "^1.0.0",
     "grunt-benchmark": "^1.0.0",
     "grunt-cli": "^1.0.0",
-    "grunt-contrib-clean": "^1.0.0 ",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-requirejs": "^1.0.0",
-    "grunt-contrib-uglify": "^2.3.0",
-    "grunt-contrib-watch": "^1.0.0",
-    "grunt-eslint": "^19.0.0",
     "grunt-jsdoc": "^2.1.0",
-    "grunt-jsinspect": "^2.0.0",
-    "grunt-jsonlint": "^1.0.4",
-    "grunt-karma": "^2.0.0",
-    "grunt-karma-coveralls": "^2.5.4",
     "grunt-open": "^0.2.3",
-    "grunt-parallel": "^0.5.0",
-    "grunt-plato": "^1.3.0",
-    "jasmine-core": "^2.2.0",
-    "jsinspect": "^0.12.4",
-    "json-server": "^0.10.0",
-    "karma": "^1.6.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.0.0",
-    "karma-firefox-launcher": "^1.0.0",
-    "karma-jasmine": "^1.0.0",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-requirejs": "^1.1.0",
-    "karma-spec-reporter": "0.0.31",
+    "jest": "^20.0.4",
     "load-grunt-tasks": "^3.1.0",
     "phantomjs-prebuilt": "^2.1.3",
     "sinon": "^2.1.0",
     "time-grunt": "^1.1.0",
-    "uglify-js": "^2.8.21"
+    "uglify-js": "^2.8.21",
+    "watch": "^1.0.2"
   },
   "browserslist": [
     "last 2 chrome versions",
@@ -77,12 +53,10 @@
     ]
   },
   "scripts": {
-    "lint": "grunt eslint:app",
-    "lint:watch": "grunt eslint:ing",
-    "test": "npm run lint && grunt karma:coverage && npm run test:build",
-    "test:watch": "grunt karma:covering",
-    "test:build": "npm run build && node test/module",
-    "test:ci": "npm test && grunt coveralls",
+    "lint": "eslint -c ./.eslintrc.js ./lib/*.js",
+    "lint:watch": "watch 'npm run lint' ./lib",
+    "test": "jest ./test/.*.test.js",
+    "test:watch": "npm test -- --watch",
     "test:perf": "npm run build && grunt benchmark",
     "docs": "grunt jsdoc && grunt open:docs",
     "preversion": "npm test",
