+++ reports/docs/atmosphere.js.html
@@ -217,7 +217,7 @@ function temperature(i) {
  * @returns {number} rate (in K/km)
 **/
 function a(i) {
-    var a = [/* eslint-disable no-magic-numbers */
+    var lapseRates = [/* eslint-disable no-magic-numbers */
         -6.5,
         0,
         1,
@@ -239,8 +239,8 @@ function a(i) {
         1.7,
         1.1,
         0
-    ]/* eslint-enable no-magic-numbers */
-    .map(function(a) {return a === 0 ? a : (a / 1000);});
+    ];/* eslint-enable no-magic-numbers */
+    var a = lapseRates.map(function(a) {return a === 0 ? a : (a / 1000);});
     return a[i];
 }
 /**
@@ -310,7 +310,7 @@ function convertMetersPerSecondToMach(speed, altitude) {
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.3</a> on Sun Jul 16 2017 11:01:46 GMT-0500 (CDT) using the <a href="https://github.com/clenemt/docdash">docdash</a> theme.
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.3</a> on Sun Jul 16 2017 11:39:37 GMT-0500 (CDT) using the <a href="https://github.com/clenemt/docdash">docdash</a> theme.
 </footer>
 
 <script>prettyPrint();</script>
