--- test/data/modules/table.js
@@ -1,88 +1,0 @@
-/**
- * @file Table Data to be used in testing
- * @author Jason Wohlgemuth
- * @module tableData
- */
-define(function() {
-    'use strict';
-
-    return [
-        {
-            "word": "ut",
-            "longitude": 62.819529,
-            "latitude": 3.162048,
-            "select": "brown",
-            "number": 786,
-            "toggle": false,
-            "guid": "66cab517-8cd7-4b8b-a123-cf939089a2ef",
-            "index": 0,
-            "_id": "5572529604f32dc471c8dc86"
-        },
-        {
-            "word": "dolor",
-            "longitude": -7.92621,
-            "latitude": -15.640182,
-            "select": "green",
-            "number": 8,
-            "toggle": true,
-            "guid": "28d0c56c-11c4-4e41-a407-bb5855965c23",
-            "index": 1,
-            "_id": "557252969d0a8fb5cb6025dd"
-        },
-        {
-            "word": "quis",
-            "longitude": 83.697629,
-            "latitude": -82.893849,
-            "select": "green",
-            "number": 363,
-            "toggle": false,
-            "guid": "dfeac747-7c58-4aa2-aca0-98e7bc2af8c3",
-            "index": 2,
-            "_id": "557252965663346f572563be"
-        },
-        {
-            "word": "duis",
-            "longitude": 157.261517,
-            "latitude": 80.783647,
-            "select": "green",
-            "number": 717,
-            "toggle": true,
-            "guid": "34e73c27-a843-4c14-a3f0-dd7ef5bb0aca",
-            "index": 3,
-            "_id": "557252962ac9a4d71ba51368"
-        },
-        {
-            "word": "esse",
-            "longitude": -92.844275,
-            "latitude": -20.814365,
-            "select": "blue",
-            "number": 777,
-            "toggle": true,
-            "guid": "6e63dad2-9504-4f2e-9a54-0e8a713044cc",
-            "index": 4,
-            "_id": "5572529685feb15b9b633560"
-        },
-        {
-            "word": "nisi",
-            "longitude": 44.223046,
-            "latitude": 12.854167,
-            "select": "brown",
-            "number": 215,
-            "toggle": false,
-            "guid": "cf7796e7-3351-4d41-9376-361753d6f677",
-            "index": 5,
-            "_id": "55725296fac8778dbc5f6cc2"
-        },
-        {
-            "word": "ea",
-            "longitude": -33.923878,
-            "latitude": 29.64501,
-            "select": "blue",
-            "number": 699,
-            "toggle": true,
-            "guid": "627c070a-18b1-4501-be95-e96d5dfc4c0d",
-            "index": 6,
-            "_id": "5572529631f711dc72b6a523"
-        }
-    ];
-});
